% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParamExpon}
\alias{ParamExpon}
\title{Parameters of exponential model}
\format{It is adviced to keep parameters in \code{\link[base]{data.frame}}. The Parameters structure is as follows:
\describe{
  \item{b0}{fixed parameter of intercept}
  \item{b1}{fixed parameter of slope}
  \item{b2}{fixed parameter of defining the decay}
  \item{varU0}{variance of random factor U0 given to fixed parameter b0}
  \item{varU1}{variance of random factor U1 given to fixed parameter b1}
  \item{corr}{correlation between random factors U0 and U1}
  \item{varE}{the variability of the residuals}
}}
\source{
Jones BC, Nair G, Shea CD, Crainiceanu CM, Cortese IC, Reich DS. Quantification of multiple-sclerosis-related brain atrophy in two heterogeneous MRI datasets using mixed-effects modeling. Neuroimage Clin. 2013 Aug 13;3:171-9. doi: 10.1016/j.nicl.2013.08.001.
}
\usage{
ParamExpon
}
\description{
Default parameters to generate micro-panel (longitudinal) data with exponencial trend. The parameters may differ per each cluster.
The parameters of each cluster are in rows. Number of rows denotes the number of clusters.
Fixed effects are taken from Jones et al. (2013).
}
\keyword{datasets}
